backend/requirements.txt
Flask>=2.2
Flask-Login>=0.6
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
Werkzeug>=2.2
